voice-assistant/requirements.txt
numpy>=1.24.0
sounddevice>=0.4.6
openai-whisper>=20231117
torch>=2.0.0
bark>=0.0.1
pyttsx3>=2.90 